--- a/docs/tools/README.md
+++ b/docs/tools/README.md
@@ -37,10 +37,35 @@ tests/
 4. **Type Safety**: All code uses type hints and passes mypy checks
 5. **PARA Integration**: Tools align with the PARA Method structure
 
-## Tool Categories
+## Implemented Tools
+
+### 🛠️ ConfigManager (Priority: P1 - Highest Impact)
+
+**Status**: ✅ Implemented | **Coverage**: 85% | **Used by**: 9/10 commands
+
+Unified YAML configuration management with schema validation, intelligent caching, and atomic updates.
+
+**Key Features**:
+- Schema-validated configuration loading with Pydantic
+- File modification time-based cache invalidation (<10ms cached reads)
+- Atomic updates with automatic backup and rollback (<100ms writes)
+- Cross-file synchronization (projects ↔ notes, team ↔ stakeholders)
+- Specialized operations for common tasks
+
+**Usage**:
+```python
+from tools import ConfigManager
+
+mgr = ConfigManager()
+project = mgr.get_project("mobile-app-v2")  # 50 lines → 1 line
+mgr.update_project("mobile-app-v2", {"status": "completed"})
+```
+
+**Documentation**: [config_manager.md](./config_manager.md)
+
+## Planned Tools
 
 ### Configuration Management
-- **ConfigManager**: Central configuration management
 - **ProjectConfig**: Project-specific configuration
 - **LearningConfig**: Learning goals and tracking configuration
 - **TeamConfig**: Team member and stakeholder management