--- a/requirements.txt
+++ b/requirements.txt
@@ -9,5 +9,8 @@ PyYAML>=6.0
 # Frontmatter parsing for notes
 python-frontmatter>=1.0.0
 
+# Configuration validation
+pydantic>=2.0.0
+
 # Optional: Enhanced CLI interface
 click>=8.0.0