--- a/tools/__init__.py
+++ b/tools/__init__.py
@@ -1,0 +1,5 @@
+"""Tools package for PARA Method utilities."""
+
+from tools.config_manager import ConfigManager, ConfigNotFoundError, ConfigValidationError
+
+__all__ = ["ConfigManager", "ConfigNotFoundError", "ConfigValidationError"]