… docs/tools/config_manager.md
… tests/test_config_manager.py