## Summary
Implements ConfigManager - the highest impact tool (P1) for unified YAML configuration management with schema validation, intelligent caching, and atomic updates as specified in #108 (Epic #106).

## Features Implemented

### Core Operations ✅
- `load_config(path, schema, use_cache)` - Load and validate YAML configs
- `update_config(path, updates, validate, backup)` - Atomic updates with rollback
- `validate_config(path, schema)` - Schema validation with Pydantic
- `merge_configs(configs, strategy)` - Deep/shallow config merging

### Specialized Operations ✅
- `get_project(project_id)` - Type-safe project retrieval
- `update_project(project_id, updates)` - Convenient project updates
- `get_all_projects(filters)` - Filtered project queries
- `get_stakeholder(stakeholder_id)` - Stakeholder profile access
- `get_team_member(member_id)` - Team member access

### Cross-File Synchronization ✅
- `sync_project_to_notes(project_id)` - Sync projects to notes cache
- `sync_team_to_stakeholders()` - Sync team roster to stakeholder contexts

### Schema Validation ✅
- Pydantic schemas for all config files (projects, team, stakeholders, integrations, decisions)
- Type-safe enums for status, priority, and other categorical fields
- Comprehensive validation with detailed error messages

### Caching System ✅
- File modification time-based cache invalidation
- `get_cache_stats()` - Cache statistics and metrics
- `clear_cache()` - Manual cache management
- <10ms cached reads (tested)

## Impact

**Before** (Per Command):
```python
# 50+ lines of boilerplate
projects_path = Path.cwd() / ".claude" / "projects.yaml"
with open(projects_path) as f:
    config = yaml.safe_load(f)
if project_id not in config["projects"]:
    raise ValueError(f"Project not found: {project_id}")
project_data = config["projects"][project_id]
# Manual validation...
```

**After**:
```python
# 1-2 lines
mgr = ConfigManager()
project = mgr.get_project("mobile-app-v2")
```

**Complexity Reduction**: ~50 lines → 1-2 lines per command
**Files Managed**: projects.yaml, team_roster.yaml, stakeholder_contexts.yaml, integrations.yaml, decision_frameworks.yaml

## Testing

✅ **28 comprehensive tests** covering:
- Basic YAML loading
- Schema validation (valid and invalid configs)
- Caching behavior and invalidation
- Atomic updates with rollback
- Cross-file synchronization
- Error handling (missing files, invalid YAML, validation errors)
- Performance benchmarks

✅ **Coverage**: 
- config_manager.py: 85%
- schemas.py: 98%

✅ **Performance**:
- Cached reads: <10ms ✅
- Writes: <100ms ✅

```bash
pytest tests/test_config_manager.py -v
# 28 passed in 1.72s
```

## Documentation

✅ Complete user guide: [docs/tools/config_manager.md](docs/tools/config_manager.md)
- Quick start examples
- Complete API reference
- Error handling patterns
- Performance guidelines
- Migration guide for existing commands

## Files Changed

- `tools/config_manager.py` - Core ConfigManager implementation
- `tools/schemas.py` - Pydantic schemas for all config files
- `tools/__init__.py` - Package exports
- `tests/test_config_manager.py` - Comprehensive test suite
- `tests/fixtures/` - Sample test configurations
- `docs/tools/config_manager.md` - Complete documentation
- `requirements.txt` - Added pydantic>=2.0.0

## Success Criteria Met

- ✅ All core operations implemented
- ✅ Test coverage >80% (85% achieved)
- ✅ Documentation complete
- ✅ Performance: <10ms cached reads, <100ms writes
- ✅ Ready for integration into commands

## Next Steps

After merge:
1. Update `/project-status` command to use ConfigManager (immediate priority)
2. Update `/follow-up-check` command
3. Update `/decide` command
4. Update `/task` command

Then proceed to #TBD (DataCollector implementation)

Closes #108

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropic.com>